requirements.txt
kafka-python-ng
influxdb-client
numpy
pandas
scipy
matplotlib
